Essentially set up a has many through to better match Thorax models
Switched to proper "population" for measure

Reverted to Rest Serializer and added expected values

The Active Model Adapter was causing some issues with the naming conventions (the all caps attributes of the populations) so I removed it and went with the Rest Serializer.  This involved changin the model names to include the original underscores, and shouldn't affect much outside the models.  The expected values are also set up.

Small fixes to the names and models